Move directories from utils to pkg
