+++ pkg/actions/connection.go
@@ -18,7 +18,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/eclipse/codewind-installer/pkg/utils/connections"
+	"github.com/eclipse/codewind-installer/pkg/connections"
 	"github.com/urfave/cli"
 )
