+++ pkg/actions/security.go
@@ -18,8 +18,8 @@ import (
 	"os"
 	"strings"
 
+	"github.com/eclipse/codewind-installer/pkg/security"
 	"github.com/eclipse/codewind-installer/pkg/utils"
-	"github.com/eclipse/codewind-installer/pkg/utils/security"
 	"github.com/urfave/cli"
 )
