+++ pkg/connections/schema.go
