+++ pkg/project/projectConnection_test.go
@@ -16,7 +16,7 @@ import (
 	"io/ioutil"
 	"testing"
 
-	"github.com/eclipse/codewind-installer/pkg/utils/connections"
+	"github.com/eclipse/codewind-installer/pkg/connections"
 	"github.com/stretchr/testify/assert"
 )
