+++ pkg/project/sync.go
@@ -25,7 +25,7 @@ import (
 	"strings"
 
 	"github.com/eclipse/codewind-installer/config"
-	"github.com/eclipse/codewind-installer/pkg/utils/connections"
+	"github.com/eclipse/codewind-installer/pkg/connections"
 	"github.com/urfave/cli"
 )
