+++ pkg/security/authenticate.go
@@ -18,8 +18,8 @@ import (
 	"net/http"
 	"strings"
 
+	"github.com/eclipse/codewind-installer/pkg/connections"
 	"github.com/eclipse/codewind-installer/pkg/utils"
-	"github.com/eclipse/codewind-installer/pkg/utils/connections"
 	"github.com/urfave/cli"
 )
